chore(clusterNode): adjust some code
